Move existing mock of API in dedicated folder
